frontend: Un-hide submit button once form fields change
When a user follows a permalink and realises the settings are not
as desired, the user should not have to wait for the slow-query
to complete before they can change the form values and re-submit.

Instead, as soon as any form field changes, un-hide the submit
button so that the user can use it to abort the query and submit
another one instead.

Change-Id: Ic750e0e43db5bd1627f4b39331f40de37d0bfbc5