frontend: Disable the submit button when loading, instead of hiding
Previously the button was hidden and the progress-bar appeared in
its place. This did mostly keep the render dimensions in-tact, given
that the bar and the button were the same height (almost).

But it still felt somewhat odd to see the button go away after pressed.

In order to be able to disable it, we have to disable it from a submit
event handler. Doing it from the click handler (which fires slightly
earlier) would result in the form never being submitted because
disabling the button from a click handler is essentially a way
of preventing the form submission.

As side-effect, using a submit handler means the manual calling
of the event handler is now redundant from the onload code, because
calling .submit() will naturally also trigger the event handler.

Change-Id: Iaf6a703b522f18b7aaba2c0c7bfa54b4cf5ee9df