Modernise layout and remove (broken) Tool Labs logo
While very abstracted, the overall layout was still heavily
inspired by MediaWiki/Wikipedia, including the "need" for a logo
on the top left corner and a sidebar.

GUC doesn't have a logo, so it always displayed the Tool Labs
logo (earlier: Toolserver). And the sidebar was otherwise empty.

The Tool Labs logo became 404 because of the rebranding to Toolforge.

Take this oppertunity to do a design refresh.

* Remove sidebar.

* Remove the page container.

* Make all components direct children of the body with shared
  styles for each of these container elements. Kind of like
  stackable panels, or cards.

* Make page-level errors into their own card,
  not nested inside the form or results.

* Fix various minor inconsistencies in padding and margin use.

* Remove unused CSS (Thanks to Chrome DevTools' Coverage feature).

* Make the styles for a grouped result similar for the case of
  the wiki having a db error and the case of the wiki being
  non-SUL.

* Generally reduce use of borders and lines.

Change-Id: I915d3df7ce7210c3217b860586cd7090209e6ccb