Consistently dereference PDO statement objects after use
In preparation for proactively closing database connections,
once we know we won't re-use them, also consistently dereference
statement objects as early as possible.

According to several online reports, closing the PDO handle and
dereferencing the PDO object is not sufficient if a statement
object is still in scope. In theory it shouldn't matter as these
should all be GC'ed and dereferenced automatically, but as a first
pass just deref'ing them manually everywhere.

Bug: T186436
Change-Id: Iee0822e723d46e2e879fbe2dfdac4f38fd5f8153