Remove the display of total edit count
Both the per-wiki and summed-together global edit count are now
removed because the queries are too slow.

There is a reason MediaWiki tracks the edit count as its own
thing associated with the user, because trying to count the whole
database of all edits ever (when only displaying the last 20) is
slow, very slow.

Before the comment and actor refactor it performed surpringly
well, really well even, but that's more a coincidence than a
design requirement, and either way is no longer the kind of query
you can do in a single web page over all wikis of all edits.

The information is still available via raw SQL queries such as
in Quarry where queries can take a bit longer, but not in GUC
for now. See T226287 for further details.

Bug: T226287
Change-Id: Ic5d64184338b104308aefd548cd04ce73e24b267