Main: Use a CentralAuth to reduce the list of wikis to check
The code for the "Big check all wikis in one UNION query" check
still exists when looking for IPs, prefixed names, prefixed IPs,
and for user names that are not known to CentralAuth for some
reason.

But, if you are searching for a single user name, and CentralAuth
knows about it, use the CentralAuth DB to decide which wikis to
check, which should make it much faster.

=== Catch

Previously, we would do one big query to check for wikis, and
then when we do the fast checks for each wiki, we only do them
for wikis where we know for sure there is at least 1 edit.

Now, we reduce the list of wikis with CentralAuth instead (if
possible). This means that if an account is auto-created somewhere
but has 0 edits, we still include it in the for-each wiki fast
check. While it should be fast, this means it will do more queries
than before if a user name has many local accounts with 0 edits.
Whether it is noticable though I'm not sure because if there are
0 edits, and the query is indexed, it typically takes about 0.03s
which would add maybe a second or two for a few hundred needless
queries. This seems worth it though, compared to making the
common case much much faster. It could potentially be avoided
by adding one big-UNION query for all CentralAuth'ed wikis at once
to look for non-zero 'editcount' fields. And then only use those.

This would mean swapping a slow "big-UNION query" over revision
tables 2 faster queries (a CentralAuth query, and a fast
"big-UNION" query over the user tables). This would still be faster
than today, but I'll first try with only the CentralAuth query.

Bug: T195515
Change-Id: I6cf21186f45a9e3cfa178abb4c86fda72d12d13c