Experiment with extra post-CentralAuth filter by editcount
Adding the big UNION back into the critical path for plain username
queries might be bad, or it mighe be good. Right now, we only filter
by existence of a local account via CentralAuth which can still
include a lot of wikis where the user has a local auto-created
account but no edits. Querying the revision table for those
with all these joins might be expensive.

Checking the user table first for non-zero edit count is faster
but the question is whether it's faster overall (given for all
wikis that *do* have edits, we would now be querying the user
table *and* the revision table).

Bug: T195515
Change-Id: Ic7549bd7de6a1d35f88549119da32910f46161b9