Query `meta_p` data explicitly from meta.labsdb instead of from s1
This effectively broke from the tools perspective. Previously, all
shards were generally co-hosted on the same underlying mysql instance,
so meta_p appeared to exist everywhere.

Now that the shards are actually separated, meta_p is only available
on s7. For now, handle this the same way as other non-wiki dbs (
such as 'centralauth'), by letting DNS tell us where to get it from
instead of assuming where it is hosted. This doesn't scale for all 900
databases, but it's okay for one or two like this.

This will automatically re-use the s7 connection when possible, just
as as centralauth today.

Bug: T282557
Change-Id: I2f73114709699bd78fda1dc16f22efedac4711b3