build: Fix PHP 8.0 and 8.1 builds, require PHP 7.3+
== Crash of lint job ==

The jobs for WMF CI were changed to include PHP 8.0 and 8.1, which
caused the repo to be unmergable due to the following fatal in php8+
from the version of php-parallel-lint used:

> Fatal error: Array and string offset access syntax with curly braces
> is no longer supported.
> in vendor/jakub-onderka/php-parallel-lint/src/Settings.php on line 145

== Crash of toolbase ==

The jobs then failed for the folllowing code in krinkle/toollabs-base
where a subclass of PDO doesn't specify the retroactively-introduced
types on the native PDO class, which coincidentally are impossible
to satisfy in a way that passes on both PHP 7.3 and PHP 8.1 unless
we turn a bunch of things off in Phan, so upstream now does.

On PHP 8.1:

> PHP Deprecated:  Return type of LoggedPDO::prepare($statement,
> $driver_options = null) should either be compatible with
> PDO::prepare(string $query, array $options = []): PDOStatement|false, or
> the #[\ReturnTypeWillChange] attribute should be used to temporarily
> suppress the notice.
> in vendor/krinkle/toollabs-base/src/LabsDB.php on line 248

And PHP 7.4/3, after meeting PHP 8's expectation:

> src/LoggedPDO.php:15 PhanParamSignatureRealMismatchHasParamTypeInternal
> Declaration of function prepare(string $query, array $options = []) should
> be compatible with internal function prepare($statement, $options = null)
> (parameter #1 has type 'string' in the real signature which cannot replace
> original parameter with no type in the real signature)

Making that and the rest of toolbase pass on PHP 8.1,
in <https://github.com/Krinkle/toollabs-base/commit/323174143dc>.

== Crash of phpcs ==

On PHP 8.1:

> PHP Fatal error:  Array and string offset access syntax with curly braces
> is no longer supported.
> in vendor/squizlabs/php_codesniffer/CodeSniffer/CLI.php on line 453

== Cash of PHPUnit ==

> There were 8 errors:
> Error: Call to undefined method ReflectionUnionType::getName()
> …guc/tests/AppTest.php:12

Reported uptream with
<https://github.com/sebastianbergmann/phpunit/issues/4543>.

Workaround for now by allowing either PHPUnit 8 or 9.

Change-Id: Id9d7be330689376c3403a84160c865b29cc8891e